[#157203061] converted menuController to a class
