+++ server/controllers/mealController.js
@@ -1,6 +1,6 @@
 import meals from '../dummy-models/meals';
 
-class MealController{
+class MealController {
   static createMeal(req, res) {
     const {
       title,
@@ -8,14 +8,14 @@ class MealController{
       price,
       img
     } = req.body;
-  
+
     if (!title || !description || !price) {
       return res.status(400).json({
         status: 'error',
         message: 'Parameters supplied incorrectly'
       });
     }
-  
+
     meals.create({
       userId: 1,
       title,
@@ -23,36 +23,36 @@ class MealController{
       price,
       img
     });
-  
+
     return res.status(201).json({
       status: 'success',
       message: 'Meal created successfully'
     });
-  };
+  }
 
   static deleteMeal(req, res) {
     const { mealId } = req.params;
-  
+
     const deleted = meals.delete(meal => parseInt(meal.id, 10) === parseInt(mealId, 10));
-  
+
     if (deleted) {
       return res.status(202).json({
         status: 'success',
         message: 'Meal deleted successfully'
       });
     }
-  
+
     return res.status(404).json({
       status: 'error',
       message: 'Meal not found'
     });
-  };
+  }
 
   static getMeal(req, res) {
     const { mealId } = req.params;
-  
+
     const foundMeal = meals.find(meal => parseInt(meal.id, 10) === parseInt(mealId, 10));
-  
+
     if (!foundMeal) {
       return res.status(404).json({
         status: 'error',
@@ -63,18 +63,18 @@ class MealController{
       status: 'success',
       data: foundMeal
     });
-  };
+  }
 
-  static getMeals(req, res){
-    return  res.status(200).json({
+  static getMeals(req, res) {
+    return res.status(200).json({
       status: 'success',
       data: meals.data
-    })
-  };
+    });
+  }
 
   static updateMeal(req, res) {
     const { mealId } = req.params;
-  
+
     let wrongKeys = false;
     Object.keys(req.body).forEach((key) => {
       if (key !== 'title'
@@ -84,31 +84,31 @@ class MealController{
         wrongKeys = true;
       }
     });
-  
+
     if (wrongKeys) {
       return res.status(400).json({
         status: 'error',
         message: 'Wrong parameters supplied'
       });
     }
-  
+
     const updatedMeal = meals.update(
       req.body,
       meal => parseInt(meal.id, 10) === parseInt(mealId, 10)
     );
-  
+
     if (Object.keys(updatedMeal).length > 0) {
       return res.status(202).json({
         status: 'success',
         message: 'Meal updated successfully'
       });
     }
-  
+
     return res.status(404).json({
       status: 'error',
       message: 'Meal not found'
     });
-  };
+  }
 }
 
-export default MealController
+export default MealController;
