--- server/controllers/mealControllerModular/index.js
@@ -1,15 +1,0 @@
-import createMeal from './createMeal';
-import getMeals from './getMeals';
-import getMeal from './getMeal';
-import updateMeal from './updateMeal';
-import deleteMeal from './deleteMeal';
-
-const MealController = {
-  createMeal,
-  getMeals,
-  getMeal,
-  updateMeal,
-  deleteMeal
-};
-
-export default MealController;
