--- server/controllers/menuController/index.js
@@ -1,17 +1,0 @@
-import createMenu from './createMenu';
-import getMenus from './getMenus';
-import getTodaysMenu from './getTodaysMenu';
-import getSpecificDayMenu from './getSpecificDayMenu';
-import updateMenu from './updateMenu';
-import deleteMenu from './deleteMenu';
-
-const MenuController = {
-  createMenu,
-  getMenus,
-  getTodaysMenu,
-  getSpecificDayMenu,
-  updateMenu,
-  deleteMenu
-};
-
-export default MenuController;
